# backend/docs/tesfafund-openapi.yaml
...
openapi: 3.0.0
info:
  version: 0.1.0
  title: TesfaFund API
  description: API specification for the TesfaFund platform.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
servers:
  - url: http://localhost:4000
...
paths:
...
  /supervisors/{id}:
    # DONE
    get:
      tags:
        - Supervisors
      summary: Retrieves the supervisor by ID, works only for authenticated supervisor.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique id of the supervisor.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Sucessful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supervisor"
        "400":
          description: Bad request - Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Recipient with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # DONE
    put:
      tags:
        - Supervisors
      summary: Updates the supervisor's details, works only for authenticated supervisor.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique id of the supervisor.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Supervisor"
      responses:
        "204":
          description: Successful update - No content to return.
        "400":
          description: Bad request - Invalid request body or parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Recipient with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
  # Permissions differ by method
  /recipients:
    # TODO: Possible differences with implementation, needs to be updated
    # (Role: Anon | Recipient | Supervisor, Owner: N/A) => Full access
    # DONE
    post:
      tags:
        - Recipients
      summary: Creates a new recipient.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                recipient:
                  type: object
                  $ref: "#/components/schemas/Recipient"
                profilePicture:
                  type: string
                  format: binary
              required:
                - recipientData
                - profilePicture
      responses:
        "201":
          description: Recipient created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipient"
        "400":
          description: Bad request - Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Conflict - Phone number already in use.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: The authentication service is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Partial access (sensitive fields not vsisible)
    # (Role: Anon | Recipient, Owner: No) => Partial access (sensitive fields not visible)
    # DONE
    get:
      tags:
        - Recipients
      summary: Retrieves recipients using filter params, hides data and parameters based on user's authorization level.
      parameters:
        - name: name
          in: query
          description: Filter by recipient's name (applies to first, middle, or last name).
          schema:
            type: string
        - name: minBirthDate
          description: The minimum birth date of the recipient (inclusive). This option is restricted to authorized supervisors.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxBirthDate
          description: The maximum birth date of the recipient (inclusive). This option is restricted to authorized supervisors.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: email
          in: query
          schema:
            type: string
            format: email
        - name: phoneNo
          description: The phone number of the recipient. This option is restricted to supervisors.
          in: query
          schema:
            type: string
        - name: page
          description: Page number for pagination. Must be a positive integer.
          in: query
          schema:
            type: integer
        - name: limit
          description: Maximum number of items per page. Must be a positive integer.
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful retrieval
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedList"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Campaign"
        "400":
          description: Bad request - Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No recipients found matching the criteria (also used instead of 403 forbidden).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
  # Permissions differ by method
  /recipients/{id}:
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => Partial access (selected fields not visible)
    # DONE
    get:
      tags:
        - Recipients
      summary: Retrieves the recipient by ID, based on user's authorization permissions.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique id of the recipient.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipient"
        "400":
          description: Bad request - Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Recipient with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # (Role: Supervisor, Owner: No) => No access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => No access
    # DONE
    put:
      tags:
        - Recipients
      summary: Updates an existing recipient.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique id of the recipient to update.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Recipient"
      responses:
        "204":
          description: Successful update - No content to return.
        "400":
          description: Bad request - Invalid request body or parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden - User is not allowed access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Recipient with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => No access
    # DONE
    delete:
      tags:
        - Recipients
      summary: Deletes a Recipient by ID, works only for authenticated Recipient.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique id of the recipient to update.
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Successful update - No content to return.
        "400":
          description: Bad request - Invalid request body or parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Recipient with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
  # Permissions differ by method
...
components:
  schemas:
...
    PaginatedList:
      type: object
      properties:
        page-number:
          type: integer
        page-count:
          type: integer
      required:
        - page-number
        - page-count
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
        phoneNo:
          type: string
      required:
        - firstName
        - middleName
        - lastName
        - dateOfBirth
        - phoneNo
    Recipient:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            email:
              type: string
              format: email
            bio:
              type: string
            profilePictureUrl:
              type: string
            socialMediaHandles:
              type: array
              items:
                $ref: "#/components/schemas/SocialMediaHandle"
          required:
            - bio
    SocialMediaHandle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        socialMediaHandle:
          type: string
      required:
        - socialMediaHandle
    Supervisor:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            email:
              type: string
      required:
        - email
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        fundraisingGoal:
          type: string
        status:
          $ref: "#/components/schemas/CampaignStatus"
        category:
          type: string
        paymentInfo:
          $ref: "#/components/schemas/PaymentInfo"
        submissionDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
        verificationDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        denialDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        launchDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        endDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        documentUrls:
          type: array
          items:
            type: string
        redactedDocumentUrls:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - fundraisingGoal
        - paymentInfo
        - category
        - endDate
    CampaignStatus:
      type: string
      enum:
        - PENDING_REVIEW
        - VERIFIED
        - DENIED
        - LIVE
        - PAUSED
        - COMPLETED
...
    PaymentInfo:
      type: object
      properties:
        paymentMethod:
          type: string
          description: The payment method used.
        phoneNo:
          type: string
          description: The phone number associated with the payment.
        bankAccountNo:
          type: string
          description: The bank account number for bank transfers.
        bankName:
          type: string
          description: The name of the bank.
      required:
        - paymentMethod
      anyOf:
        - required:
            - phoneNo
        - required:
            - bankAccountNo
            - bankName
...
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
      additionalProperties: true
  securitySchemes:
    bearerHttpAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token using a JWT